+++ CHANGELOG.rst
@@ -4,9 +4,12 @@ Changelog
 Unreleased
 ----------
 
+- (Added) ``sismic-behave`` CLI now accepts a ``--steps`` parameter, which is a list of filepaths containing the steps
+  implementation.
 - (Changed) Reorganisation of ``docs/examples``.
 - (Fixed) Coverage data for ``sismic-behave`` takes the initialization step into account (regression
   introduced in 0.21.0).
+
 
 
 0.22.3 (2016-07-06)